+++ index.js
@@ -2,7 +2,6 @@ require('dotenv').config()
 
 const server = require('./api/server');
 
-
 const port = process.env.PORT || 8888;
 
 server.listen(port, () => {
